+++ colossalai/pipeline/middleware/adaptor/__init__.py
@@ -1,0 +1,3 @@
+from .fx import get_topology as get_fx_topology
+
+__all__ = ['get_fx_topology']
