--- logs/partition1.txt
@@ -1,14 +1,0 @@
-GraphModule(
-  (layers_3): Linear(in_features=10, out_features=10, bias=False)
-  (layers_4): Linear(in_features=10, out_features=10, bias=False)
-  (layers_5): Linear(in_features=10, out_features=10, bias=False)
-)
-
-
-
-def forward(self, layers_2):
-    layers_3 = self.layers_3(layers_2);  layers_2 = None
-    layers_4 = self.layers_4(layers_3);  layers_3 = None
-    layers_5 = self.layers_5(layers_4);  layers_4 = None
-    return layers_5
-
