--- logs/script
@@ -1,25 +1,0 @@
-if pp_rank == 0:
-        with open('logs/modules.txt', 'w') as f:
-            f.write(str(annotated_model))
-            
-        with open('logs/tracer', 'w') as f:
-            for node in gm.graph.nodes:
-                f.write(f'node {node} | args {node.args} | users {node.users}\n-----\n')
-        with open('logs/top_mod.txt', 'w') as f:
-            f.write(str(split_model))
-        
-if pp_rank == 0:
-        with open('logs/partition0.txt', 'w') as f:
-            f.write(str(partition))
-            
-    if pp_rank == 1:
-        with open('logs/partition1.txt', 'w') as f:
-            f.write(str(partition))
-            
-    if pp_rank == 2:
-        with open('logs/partition2.txt', 'w') as f:
-            f.write(str(partition))
-            
-    if pp_rank == 3:
-        with open('logs/partition3.txt', 'w') as f:
-            f.write(str(partition))
