--- logs/top_mod.txt
@@ -1,12 +1,0 @@
-GraphModule(
-  (submod_0): GraphModule()
-  (submod_1): GraphModule()
-)
-
-
-
-def forward(self, x : torch.Tensor):
-    submod_0 = self.submod_0()
-    submod_1 = self.submod_1(x, submod_0);  x = submod_0 = None
-    return submod_1
-
