--- logs/tracer
@@ -1,54 +1,0 @@
-node x | args () | users {linear: None}
------
-node layers_0_weight | args () | users {linear: None}
------
-node pipe_split | args () | users {}
------
-node layers_0_bias | args () | users {add: None}
------
-node linear | args (x, layers_0_weight) | users {add: None}
------
-node add | args (linear, layers_0_bias) | users {linear_1: None}
------
-node layers_1_weight | args () | users {linear_1: None}
------
-node layers_1_bias | args () | users {add_1: None}
------
-node linear_1 | args (add, layers_1_weight) | users {add_1: None}
------
-node add_1 | args (linear_1, layers_1_bias) | users {linear_2: None}
------
-node layers_2_weight | args () | users {linear_2: None}
------
-node layers_2_bias | args () | users {add_2: None}
------
-node linear_2 | args (add_1, layers_2_weight) | users {add_2: None}
------
-node add_2 | args (linear_2, layers_2_bias) | users {linear_3: None}
------
-node layers_3_weight | args () | users {linear_3: None}
------
-node layers_3_bias | args () | users {add_3: None}
------
-node linear_3 | args (add_2, layers_3_weight) | users {add_3: None}
------
-node add_3 | args (linear_3, layers_3_bias) | users {linear_4: None}
------
-node layers_4_weight | args () | users {linear_4: None}
------
-node layers_4_bias | args () | users {add_4: None}
------
-node linear_4 | args (add_3, layers_4_weight) | users {add_4: None}
------
-node add_4 | args (linear_4, layers_4_bias) | users {linear_5: None}
------
-node layers_5_weight | args () | users {linear_5: None}
------
-node layers_5_bias | args () | users {add_5: None}
------
-node linear_5 | args (add_4, layers_5_weight) | users {add_5: None}
------
-node add_5 | args (linear_5, layers_5_bias) | users {output: None}
------
-node output | args (add_5,) | users {}
------
